Replace non-standard uint with standard C++ types for GCC 15 compatibility
GCC 15 does not implicitly provide the POSIX uint typedef, and its new
-Wtemplate-body flag (enabled by -Wall) catches uses inside uninstantiated
templates that older GCC versions silently ignored.

Replace uint with size_t for loop indices and unsigned int for unused
Kokkos lambda parameters. Also guard the Clang-only -Wno-nan-infinity-disabled
flag in CMakeLists.txt behind a compiler check.

+++ src/include/TempLat/lattice/algebra/random/randomgaussianfield.h
@@ -115,7 +115,7 @@ namespace TempLat
           global_coord, hermitianPartner);
 
       // We do not need coordinates actually, but rather (positive!) global indices.
-      for (uint d = 0; d < NDim; ++d) {
+      for (size_t d = 0; d < NDim; ++d) {
         if (global_coord[d] < 0) global_coord[d] += mGlobalSizes[d];
         if (hermitianPartner[d] < 0) hermitianPartner[d] += mGlobalSizes[d];
       }
